services/agents/requirements.txt
# AI/ML
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5

# Data processing
pandas==2.1.4
numpy==1.26.3

# Document generation
python-docx==0.8.11
reportlab==4.0.7

# HTTP & Async
aiohttp==3.9.1
httpx>=0.24,<0.26  # Compatible with supabase 2.3.3
requests>=2.31.0

# Data formats
pyyaml==6.0.1

# Database
supabase==2.3.3

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3

# Web Push Notifications
pywebpush>=1.14.0,<2.0.0  # <2.0 for httpx 0.25 compatibility
py-vapid>=1.9.0
